Minor code refactor for App/AppConfig
